hwid/service: make the API protobuf definition match the style guide
In the past the request and response payload of the HWID service
is in form of JSON.  While migrating to the new stubby endpoint,
the protobuf definition was introduced.  And in order to keep the
field names backward-compatible, the field names in the protobuf
deliberately followed `camelCase` but not the suggested `snake_case`.

Now since the JSON over HTTP endpoint is removed, this CL safely
rename the fields in the protobuf files so that it follows the
suggested naming rule.

BUG=None
TEST=./deploy/cros_hwid_service.sh test

Change-Id: I5932349df3dcb78966490426bd71f4c7bb00ef49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2856058
Reviewed-by: Clark Chung <ckclark@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>