image_tool: Verify cros_config between test/release images
We used to use `gooftool verify_cros_config` to check if the values
we use in the factory test are the same as the values that are in
the release image.

However, `gooftool verify_cros_config` is in the finalization step and
it is the last test that the users verify. They need to re-test all
tests if they find the config is not synced. This CL adds a check at the
bundle/image creation stage.

BUG=b:181089724
TEST=(0) BUNDLE=/tmp/bundle && BOARD=xxx && PROJECT=yyy
     (1) rm -rf ${BUNDLE} && mkdir -p ${BUNDLE}
     (2) printf "
board: ${BOARD}
project: ${PROJECT}
designs: boxster_designs
bundle_name: 20210505_pvt
server_url: http://192.168.71.101:8080/
firmware: release_image
hwid: real
release_image: 13949.0.0
test_image: 13600.177.0
toolkit: 13600.177.0" > ${BUNDLE}/MANIFEST.yaml
     (3) bin/finalize_bundle ${BUNDLE}/{MANIFEST.yaml,} --jobs 8
     (4) BUNDLEX=${BUNDLE}/factory_bundle_${PROJECT}_20210505_pvt
     (5) cp -f py/tools/image_tool.py ${BUNDLEX}/setup
     (6) pushd ${BUNDLEX}
     (7) setup/image_tool.py rma create -o ./rma.bin \
           --project ${PROJECT}
     (8) popd

Change-Id: I243d0348857fe2e6bdc53268c86c6271f56d1b1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2875005
Reviewed-by: Ting-Hsuan Wang <phoebewang@chromium.org>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>