thunderbolt_loopback: Add retimer firmware check
Partner mentioned, that they will be flashing firmware of certain
version for the Burnside Bridge Retimer on Mass Production / shipping
devices.

We add a check to the factory test process to ensure that the retimer
firmware is actually the above mentioned version.

The retimer firmware version is readable from sysfs.

BUG=b:181360981, b:187264683
TEST=run FFT.TBTLoopback01
TEST=run FFT.TBTLoopback03

Change-Id: Ic71743470807e27091fa700f4059bfae5d89ca2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2886582
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Ting-Hsuan Wang <phoebewang@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>