camera: add cros_config as default method to search for device index
Originally, the device index of the camera is obtained by checking
the camera_characteristics.conf. However, the camera team is going
to move the facing setting to cros_config. We thus add cros_config
as a default method to read the facing setting. We also keep the
checking of camera_characteristics.conf as a fall back approach.

BUG=b:186497201
TEST=test manually on DUT

Change-Id: I65c9561aeff653ec0323c9f56c52a89a0f6e0da1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2899830
Tested-by: Ting-Hsuan Wang <phoebewang@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Ting-Hsuan Wang <phoebewang@chromium.org>