gooftool: Clear AP-RO hash in cr50 before setting it
If the AP-RO hash is already set in cr50, we cannot set it again. This
can happen if the finalization process is interrupted, so we set the
AP-RO hash but didn't set the board ID. Because the AP-RO hash can still
be cleared before the board ID is set, we should always try to clear the
AP-RO hash before we set it.

BUG=b:188567171
TEST=Verified by the partner

Change-Id: I4e1e0805091cb5bc96f89af2f90e3b35e93b0489
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2909703
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>