camera: detect whether the camera is well assembled
Sometimes camera may be badly assembled and the images captured will
have black edges. We calculate the luminance of the boundary region
of the image. If it is lower than or equal to a predefined threshold,
we consider the image containing black edges.

We also add a new function to draw the detection region of the QR
code to avoid user placing QR code at the boundary or center regions.

BUG=b:188002711
TEST=manually tested on DUTs

Change-Id: I6f3220ee697beff87fd2f9e8484216a4ae992a9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2987608
Tested-by: Ting-Hsuan Wang <phoebewang@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Ting-Hsuan Wang <phoebewang@chromium.org>