tablet_rotation: isInTabletPhysicalState is renamed
The property isInTabletPhysicalState is changed to isAutoRotationAllowed
in https://crrev.com/c/2880971.

Note that the definition is changed slightly:
  Auto rotation is allowed when the device is in a physical tablet state
  or kSupportsClamshellAutoRotation is set.

So, if kSupportsClamshellAutoRotation is set, the value would be true
even if the tablet mode switch is off (false).

But this should be fine, because we will run the "tablet_mode" test
before running this test, and the test will check the "tablet mode
switch" in evtest.

BUG=b:190413571
TEST=manually on device

Change-Id: I3cf556604694a9d221a4318571206b075d48040b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2952407
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Ting-Hsuan Wang <phoebewang@chromium.org>
Tested-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Stimim Chen <stimim@chromium.org>