pytests: audio_loop: add arguments for input gain
Even if we have `cras_enabled=True` in the argument, the 'audiofun' test
mode would still bypass CRAS.  This is intended because we might want to
test specific channel, which is hard if we are using CRAS.

However, CRAS would adjust the GAIN of input devices, try to balance the
value of "IntrinsicSensitivity" set in HiFi.conf.  For example, if the
value is set to "-3200", then it is -32dB, and CRAS would try to
adjust it to -6dB ==> CRAS will apply a 26dB gain.  And you will see
something like the following in the output of `cras_test_client
--dump_server_info`:

  Stable Id        ID     Gain  UI       Plugged ...
  (2a118f1d)      9:0        26 1.000000     yes ...

This CL would attach a `gain X` argument to the recorder command of
audiofuntest. To apply a 26dB gain, set `tests_to_conduct` as:

  {
    "type": "audiofun",
    "input_gain": 26,
    ...
  }

BUG=b:190558228
TEST=manual test on device

Change-Id: I0b388d07a08b18a4bf3cb5a2878db7355ef2ccb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2963793
Tested-by: Stimim Chen <stimim@chromium.org>
Reviewed-by: Gene Chang <genechang@google.com>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Stimim Chen <stimim@chromium.org>