Instalog: output_factory_report: Parse serial numbers again if we failed
The serial_numbers field in Eventlog is added from CL:362730. For the
projects before this CL, their reports don't have the field, so we need
to parse serial numbers by other method. However, we didn't find a
common event or field with serial numbers on these reports, so we use
the regular expression to parse.

BUG=b:189768657
TEST=manually test

Change-Id: I5269f64e627ec7a97e857161c5602042e80530c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2972885
Commit-Queue: Chun-Tsen Kuo <chuntsen@chromium.org>
Tested-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>