gooftool: Check the rootkey between release and running firmware
To prevent wrong firmware flashed or wrong firmware rootkey written. we
should compare the rootkey from the running firmware and the rootkey
from the firmware updater in the release rootfs. If both rootkeys are
the same, It means that the firmware version is correct. Otherwise,
the wrong firmware version is flashed, and the finalization should not
be passed.

BUG=b:189806435
TEST=manually test on DUT

Change-Id: I453918d149e875cdb67bff1b5045464b0239489e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2987612
Tested-by: Gene Chang <genechang@google.com>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Gene Chang <genechang@google.com>