pytests: Add check_retimer_firmware to check retimer firmware
We split this test from thunderbolt_loopback.py since the behavior of
ToT image is changed and checking the retimer firmware does not
require a loopback card now.

One more thing is the retimer devices do not get enumerated if anything
except chargers connected to the port before boot so we need to make
more steps to ensure this condition.

BUG=b:181360981
TEST=run CheckRetimerFirmwareGroup on devices

Change-Id: I68d2f61b44270f26015dc631c24dccbc2204f8f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2994487
Reviewed-by: Ting-Hsuan Wang <phoebewang@chromium.org>
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>