cbi_utils: invalidate cache when getting cbi values
EC would cache the CBI value.  And even if the set command failed, the
cache is still updated with the new value.  Currently we rely on the
return code of the set command to determine the successfulness, so it's
not a problem.

But we should always invalidate the cache when getting CBI values to
avoid confusions.

BUG=None
TEST=make test

Change-Id: Ib50997b5198670fd07a37ea1f1e16adf08a07f06
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2999452
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>