camera: conduct camera_assemble and qr code test together
We add a new test which conducts camera_assemble and qr code test
simultaneously. By conducting camera_assemble alone may result in
unintended results since it relies on the luminance of boundary and
center region. For example, if the operator wears black clothes, the
luminance of center region would be low and thus the test might pass.
By running these two test together, we could ensure that the operator
has used a white paper with QR code on it to cover the FOV of the
camera. Hence, this will reduce the false positive rate.

BUG=b:188002711
TEST=manually tested on DUT

Change-Id: I1ca5a0be70af0b547712d0393f9c768f0984a17c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3010481
Tested-by: Ting-Hsuan Wang <phoebewang@chromium.org>
Reviewed-by: Chung-Sheng Wu <chungsheng@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Ting-Hsuan Wang <phoebewang@chromium.org>