hwid/v3: adds error codes into the return of ContentsAnalyzer's methods
Previously the response data of methods in `contents_analyzer` only
records a list of string of hint messages if some errors are detected
in the given HWID DB payload.  This CL revises the data structure to
record predefined error codes that indicates the type of the error.

In addition, in HWID service side, this CL revises the response
format of `ValidateHwidDbEditableSectionChange` RPC to record
a list of error messages instead of a concatenated one.

BUG=b:193293227
TEST=make test; ./deploy/cros_hwid_service.sh test

Change-Id: I74a9cb4c5d0db2d8b0c4f41cbd1cf4b3cac4f81b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3023754
Reviewed-by: Clark Chung <ckclark@chromium.org>
Tested-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>