goofy: support modifier in bindKey and change screenshot hotkey
Originally, the screenshot hotkey is bind to `S`. However, this
setting conflicts with the start state in "AudioJack" test. We thus
add support to modifier in bindKey and modify the screenshot hotkey
to `Ctrl + S`. Moreover, we normally don't bind keys with modifier in
pytests, so adding a modifier might be safer to prevent collision.
Also, we store the screenshots under /var/factory/log/screenshots, so
that the screenshots will be zipped when using factory_bug.

BUG=b:188315163
TEST=test manually on DUT

Change-Id: I1fd560309cdf4dc31364d01b7340714199aea425
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3071509
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Tested-by: Ting-Hsuan Wang <phoebewang@chromium.org>
Commit-Queue: Ting-Hsuan Wang <phoebewang@chromium.org>