cros_board_utils: remove board check and fuzzy board searching
cros_board_utils checks the validity of the input board by comparing
with the boards listed in chromiumos-overlay/eclass/cros-board.eclass.
However, if the board we would like to use is removed or not added in
the cros-board.eclass file, the check would fail. For instance, in
this CL (https://crrev.com/c/3057394), some of the obsolete boards are
removed and board `daisy_spring` is one of them. gsutil_unittest fails
since cros_board_utils can not find `daisy_spring`. We thus remove the
board check and let the test fails if the input board is not valid.
We also remove fuzzy board searching (e.g. board `spring` means
`daisy_spring`), since it relies on the boards listed in
cros-board.eclass.

BUG=b:195382786
TEST=make test

Change-Id: I69bf286ee6b3c3ffd7bc60c3860e312641913e9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3068906
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Ting-Hsuan Wang <phoebewang@chromium.org>
Commit-Queue: Ting-Hsuan Wang <phoebewang@chromium.org>