hwid/service: stop setting is_vp_related in component if unnecessary
The field is_vp_related is used to determine if the component should be
reported in GetDutLabels API.  However, the method call
GetAllComponentVerificationPayloadPieces takes up the majority of the
response time which is unnecessary.  This CL adds a flag
`require_vp_info` to indicate if the vp info is required.

BUG=b:194462833
TEST=(FACTORY REPO)$ ./deploy/cros_hwid_service.sh test

Change-Id: Ica3aaa24b211f75dd2735995885c4a574050a813
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3071512
Tested-by: Clark Chung <ckclark@chromium.org>
Auto-Submit: Clark Chung <ckclark@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>