image_tool: remove partition 10 if it is the last partition
In disk_layout_v3, the NBR team put MiniOS-B at the end of the disk.
Since now MiniOS-B is the last partition, we cannot expand or shrink
the size of stateful partition. To solve this, we remove partition 10,
which contains MiniOS-B, from the partition table of the factory
install shim.

BUG=b:198542742, b:197919353
TEST=(1) ./setup/image_tool rma create -o brya_rma.bin.
     (2) Install the RMA shim on brya DUT.
     (3) Verify that DUT works normally.

Change-Id: I701069c096ad7e1034c78893af905902e9a666f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3139582
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Ting-Hsuan Wang <phoebewang@chromium.org>
Commit-Queue: Ting-Hsuan Wang <phoebewang@chromium.org>