partition_table: remove MINIOS-B and add it back after resizing
partition_table pytest expands the stateful partition to the end of
the disk. However, in disk_layout_v3, the last partition is MINIOS-B,
which prevents us from expanding the size of the stateful partition.
To solve this, we first remove partition MINIOS-B, expand the size of
the stateful partition, add MINIOS-B back by copying the content from
MINIOS-A.

BUG=b:199117959, b:197919353
TEST=(1) Resize the stateful partition on partition table such that
         the size is less than 95% disk size.
         (Using `cgpt add -i 1 -s [size] /dev/nvme0n1`)
     (2) Run partition_table test
     (3) Check if stateful partition is resized to desired size and
         check if MINIOS-B is at the last partition.
	 (Using `cgpt show /dev/nvme0n1`)

Change-Id: I81e35fecdb835b4e7539790be885e0f0b9a28f2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3144772
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Tested-by: Ting-Hsuan Wang <phoebewang@chromium.org>
Commit-Queue: Ting-Hsuan Wang <phoebewang@chromium.org>